fix: Switch to Scriptable Build Pipeline (#43)
All material keywords are not saved in ABs
Unfortunately, it's another [nasty Unity
bug](https://issuetracker.unity3d.com/issues/assetbundles-shader-keyword-is-lost-in-build-when-shader-is-in-different-assetbundle-than-material)
:facepalm: Emission and Normal map keywords are not saved either.
SHADER KEYWORD IS LOST IN BUILD WHEN SHADER IS IN DIFFERENT ASSETBUNDLE
THAN MATERIAL
Perfectly describes the situation.
The reason why it works in the old renderer is because they do this
manually:
public static void OptimizeMaterial(Material material)
        {
//NOTE(Brian): Just enable these keywords so the SRP batcher batches
more stuff.
            material.EnableKeyword("_EMISSION");
            material.EnableKeyword("_NORMALMAP");

if (!material.IsKeywordEnabled("_ALPHATEST_ON") &&
material.HasProperty(ShaderUtils.Cutoff))
                material.SetFloat(ShaderUtils.Cutoff, 0);

            material.EnableKeyword("_ALPHATEST_ON");
It's a pure hack :melting_face: You never should call anything like this
for bundled materials. Also in advance, they of course gather all these
materials from loaded Asset Bundles... And I am not going to introduce
this weirdness in Explorer Alpha.
The solution to it is using a [new build
pipeline](https://docs.unity3d.com/Packages/com.unity.scriptablebuildpipeline@1.21/manual/index.html)
I already modified Asset Bundle Pipeline and it works perfectly with one
exception: you can't cross-reference asset bundles generated in old and
new ways. E.g. we have a common bundle with the shader, and the old one
does not work with newly generated bundles. Most probably because GUIDs
are differently generated.
Though, it is still possible to load new ones by the old logic so
changes to the client are not needed
I also found a big red flag with our way of generating metadata.json :
we run the build twice, the second time forcing a rebuild without
caching.
With Scriptable Build Pipeline clearing cache is no longer needed as
cache works absolutely differently and is created per asset, not per
asset bundle. So it can be reused even if the content of the same asset
bundles is reshuffled.
The second bonus is the general performance of SRP which is considerably
better.
So after all we will be able to convert Asset Bundles much faster.

- Scriptable Build Pipeline integrated
- `pipeline` CLI option set to `Scriptable` by default, it's possible to
change it to `default`

---------

Co-authored-by: Juan Molteni <juanignaciomolteni@gmail.com>