fix: Avoid duplicated scene entities (#221)
For MVMF 2025, there was a stage at `-62,-64`. The scene lod manifest
builder creating duplicated entities of the `GLTFContainer` entity; as
exemplified here.


[StaticSceneDescriptor_broken.json](https://github.com/user-attachments/files/23460334/StaticSceneDescriptor_broken.json)


[bafkreic5xdahovfi4cigvjtdetx6cgeadcfioyffiqh26kispclmeknuuy-lod-manifest.json](https://github.com/user-attachments/files/23460399/bafkreic5xdahovfi4cigvjtdetx6cgeadcfioyffiqh26kispclmeknuuy-lod-manifest.json)


To mitigate the situation, we create a hash set of entity as tracker,
instead of a list. Therefore, the same entity cannot be twice.

This would need to be solved on the manifest builder level