feat: social emotes shape: storing the start pose of the hips of the avatar in outcome animations (#223)
In order to move the avatar that reacts to a social emote to the very
first position in which it should be when the outcome animation starts,
it was necessary to get that information (the position and rotation of
the hips of the avatar) upfront, so it can be used in runtime by the
explorer. There is one entry per outcome animation.

---------

Signed-off-by: Juan Ignacio Molteni <juanignaciomolteni@gmail.com>
Co-authored-by: Juan Ignacio Molteni <juanignaciomolteni@gmail.com>